add cqt eval

+++ src/lib/data_wrangler.py
@@ -11,7 +11,7 @@ import pickle
 
 
 SPLITS_DIR = 'splits/'
-D_TYPE = 'float16'
+D_TYPE = 'float32'
 
 
 def save_training_results(history, experiment_results_dir, experiment_id, model):
